feat: complex model class cleanup: separate files for different importers
