+++ lib.flist
@@ -37,10 +37,16 @@ lib/graphics/primitives/texture.o
 lib/graphics/primitives/framebuffer.o
 lib/graphics/primitives/flat_renderer.o
 lib/graphics/primitives/render_frame.o
-lib/graphics/importers/importers.o
+
+lib/graphics/importers/materials.o
+lib/graphics/importers/models.o
+lib/graphics/importers/shaders.o
+lib/graphics/importers/textures.o
+
 lib/graphics/objects/uniform_set.o
 lib/graphics/objects/material.o
 lib/graphics/objects/model.o
+lib/graphics/objects/complex_model.o
 lib/graphics/objects/scene.o
 lib/graphics/objects/ambient_light.o
 lib/graphics/objects/point_light.o
