+++ lib/logics/components/visual/static_mesh.h
@@ -11,12 +11,11 @@
 
 #pragma once
 
-#include "graphics/objects/model.h"
+#include "graphics/objects/complex_model.h"
 #include "logics/scene_component.h"
 
 struct StaticMesh : public SceneComponent {
-    explicit StaticMesh(const Mesh& mesh, const Material& material);
-    explicit StaticMesh(const Model& model);
+    explicit StaticMesh(const ComplexModel& model);
 
     void spawn_self(Scene& scene) override;
 
@@ -25,10 +24,10 @@ struct StaticMesh : public SceneComponent {
         return model_.get_object_matrix();
     }
 
-    Model& get_model() { return model_; }
-    const Model& get_model() const { return model_; }
+    ComplexModel& get_model() { return model_; }
+    const ComplexModel& get_model() const { return model_; }
     void set_model(const Model& model) { model_ = model; }
 
    private:
-    Model model_;
+    ComplexModel model_;
 };
